Split and rename some of the XcTest execution internals
